+++ CHANGELOG.md
@@ -1,19 +1,23 @@
 # Changelog
 
 ### v0.0.36
+
 - Fix importing React issue which causes crash while implementing PeopleListItem and AnimatedHeader
 - Fix CompactCardListItem's shadow issue on IOS side
 - Fix PeopleListItem component UI
 - Add link label in Tab component
+- Add Core Milestone Component
 
 ---
 
 ### v0.0.35
+
 - Add done button in the DatePicker component for ios
 
 ---
 
 ### v0.0.34
+
 - Add LinearGradient component
 - Add CompactCardListItem component
 - Add PeopleListItem component